Use same style as other API methods
